Factor out management of wasm-specific resources into `Driver` type.
The new `Driver` type can be used to customize how go-app environment is
set up on the client side. It is responsible for pre-rendering and
registering scripts and styles that are necessary to load and start
app's business logic.

By default, a Go WebAssembly driver is assumed, which maintains the
original behavior with showing a loader progress bar and starting up a
Go wasm binary.

Users can customize the loading process and the scripts necessary for it
by providing their own Driver implementations - at their own peril.

One notable change is that app.js has been split into two parts:

 - The code responsible for managing service worker remains in app.js
   and is included regardless of the driver used.
 - The code responsible for setting up wasm was moved to driver-wasm.js
   and is used by the `goWasmDriver` implementation.